requirements.txt
flask>=3.0.2
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
pyyaml>=6.0.1
scapy>=2.5.0
gunicorn>=21.2.0
dash>=2.14.0
plotly>=5.18.0
